Switch database to PostgreSQL

+++ src/main/java/com/cobank/api/dto/TransactionHistoryResponse.java
@@ -1,12 +1,13 @@
 package com.cobank.api.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record TransactionHistoryResponse(
         String iban,
         TransactionType transactionType,
-        double amount,
-        double resultingBalance,
+        BigDecimal amount,
+        BigDecimal resultingBalance,
         LocalDateTime timestamp,
         String description
 ) {}
